requirements.txt
huggingface-hub
transformers
torch
python-dotenv
flask
fireworks-ai 